+++ vraptor-core/src/main/java/br/com/caelum/vraptor/observer/upload/UploadedFileConverter.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package br.com.caelum.vraptor.interceptor.multipart;
+package br.com.caelum.vraptor.observer.upload;
 
 
 import javax.inject.Inject;
