+++ vraptor-core/src/test/java/br/com/caelum/vraptor/observer/download/ByteArrayDownloadTest.java
@@ -1,4 +1,4 @@
-package br.com.caelum.vraptor.interceptor.download;
+package br.com.caelum.vraptor.observer.download;
 
 import static org.junit.Assert.assertArrayEquals;
 import static org.mockito.Mockito.times;
@@ -6,24 +6,20 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
 
 import javax.servlet.ServletOutputStream;
 import javax.servlet.http.HttpServletResponse;
 
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
-import br.com.caelum.vraptor.observer.download.FileDownload;
+import br.com.caelum.vraptor.observer.download.ByteArrayDownload;
 
-public class FileDownloadTest {
+public class ByteArrayDownloadTest {
 
-	private File file;
 	private byte[] bytes;
 	private @Mock HttpServletResponse response;
 	private ServletOutputStream socketStream;
@@ -34,15 +30,9 @@ public class FileDownloadTest {
 		MockitoAnnotations.initMocks(this);
 
 		bytes = new byte[] { (byte) 0 };
-		outputStream = new ByteArrayOutputStream();
+		this.outputStream = new ByteArrayOutputStream();
 
-		file = File.createTempFile("test", "vraptor");
-		
-		try (FileOutputStream fileStream = new FileOutputStream(file)) {
-			fileStream.write(bytes);
-		}
-
-		socketStream = new ServletOutputStream() {
+		this.socketStream = new ServletOutputStream() {
 			@Override
 			public void write(int b) throws IOException {
 				outputStream.write(b);
@@ -52,14 +42,9 @@ public class FileDownloadTest {
 		when(response.getOutputStream()).thenReturn(socketStream);
 	}
 
-	@After
-	public void tearDown() {
-		file.delete();
-	}
-
 	@Test
-	public void shouldFlushWholeFileToHttpResponse() throws IOException {
-		FileDownload fd = new FileDownload(file, "type");
+	public void shouldFlushWholeStreamToHttpResponse() throws IOException {
+		ByteArrayDownload fd = new ByteArrayDownload(bytes, "type", "x.txt");
 		fd.write(response);
 
 		assertArrayEquals(bytes, outputStream.toByteArray());
@@ -67,11 +52,10 @@ public class FileDownloadTest {
 
 	@Test
 	public void shouldUseHeadersToHttpResponse() throws IOException {
-		FileDownload fd = new FileDownload(file, "type", "x.txt", false);
+		ByteArrayDownload fd = new ByteArrayDownload(bytes, "type", "x.txt");
 		fd.write(response);
-		
+
 		verify(response, times(1)).setHeader("Content-type", "type");
-		verify(response, times(1)).setHeader("Content-disposition", "inline; filename=x.txt");
 		assertArrayEquals(bytes, outputStream.toByteArray());
 	}
 }
