+++ vraptor-core/src/test/java/br/com/caelum/vraptor/observer/upload/UploadedFileConverterTest.java
@@ -1,4 +1,4 @@
-package br.com.caelum.vraptor.interceptor.multipart;
+package br.com.caelum.vraptor.observer.upload;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.when;
@@ -9,6 +9,9 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+
+import br.com.caelum.vraptor.observer.upload.UploadedFile;
+import br.com.caelum.vraptor.observer.upload.UploadedFileConverter;
 
 public class UploadedFileConverterTest {
