+++ vraptor-musicjungle/src/main/java/br/com/caelum/vraptor/musicjungle/config/CustomMultipartConfig.java
@@ -3,7 +3,7 @@ package br.com.caelum.vraptor.musicjungle.config;
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Specializes;
 
-import br.com.caelum.vraptor.interceptor.multipart.DefaultMultipartConfig;
+import br.com.caelum.vraptor.observer.upload.DefaultMultipartConfig;
 
 @ApplicationScoped
 @Specializes
