+++ vraptor-musicjungle/src/main/java/br/com/caelum/vraptor/musicjungle/files/Musics.java
@@ -8,8 +8,8 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 
-import br.com.caelum.vraptor.interceptor.multipart.UploadedFile;
 import br.com.caelum.vraptor.musicjungle.model.Music;
+import br.com.caelum.vraptor.observer.upload.UploadedFile;
 
 public class Musics {
